#229 hide/show work
but the ReactFauxDom doesn’t truly clear out the elements from one
selection to the next. This is problematic b/c data is scoped according
to the display setting and passed into OrderedWordCloud. If OWC doesn’t
update the data as d3 would w/r/to enter/exit, it leaves elements there
- with uncleared and inaccurate hide/show class settings.