#427
we are doing this for a few use cases/flow patterns
sources and collections are unselected when the user clicks the icon to
return home, which means that there is a case when both sourceId and
collectionId are null. Hence, the subheader cannot infer from the
source or collection object alone what it should show. Therefore we
have to look at the url